requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.10.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
ImageHash==4.3.1
lxml==4.9.2
magic-filter==1.0.9
multidict==6.0.4
numpy==1.25.0
Pillow==9.5.0
pytz==2023.3
PyWavelets==1.4.1
requests==2.31.0
scipy==1.11.1
six==1.16.0
soupsieve==2.4.1
tzlocal==5.0.1
urllib3==2.0.3
yarl==1.9.2
